+++ src/main/java/ai/timefold/solver/benchmarks/competitive/tsplib95/TSPLIBDataset.java
@@ -181,7 +181,7 @@ public enum TSPLIBDataset implements Dataset<TSPLIBDataset> {
     }
 
     @Override
-    public int getBestKnownDistance() {
+    public double getBestKnownDistance() {
         return bestKnownDistance;
     }
